pyproject.toml
[project]
name = "spotify-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.116.2",
    "joblib>=1.5.2",
    "kagglehub>=0.3.13",
    "logfire[fastapi]>=4.9.0",
    "pandas>=2.3.2",
    "pydantic-settings>=2.10.1",
    "scikit-learn>=1.7.2",
    "streamlit>=1.49.1",
]
